requirements.txt
streamlit
networkx
matplotlib
pillow
numpy